ft: added CatalogRepository class, to handle all the catalog operations (retrieve, store, filter, save inputcat to models, write/load testing cats, etc.
